@@ -3,37 +3,42 @@
   "version": "0.0.0",
   "private": true,
   "dependencies": {
-    "@dataspecer/backend-utils": "^0.0.0",
+    "@dataspecer/backend-utils": "*",
     "@dataspecer/core": "*",
-    "@dataspecer/federated-observable-store": "^0.0.0",
-    "@dataspecer/federated-observable-store-react": "^0.0.0",
-    "@emotion/react": "^11.5.0",
-    "@emotion/styled": "^11.3.0",
-    "@mui/icons-material": "^5.1.0",
-    "@mui/lab": "^5.0.0-alpha.63",
-    "@mui/material": "^5.1.0",
-    "@mui/x-data-grid": "^5.0.0-beta.7",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@types/jest": "^26.0.15",
-    "@types/node": "^12.0.0",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
+    "@dataspecer/federated-observable-store": "*",
+    "@dataspecer/federated-observable-store-react": "*",
+    "@emotion/react": "^11.9.0",
+    "@emotion/styled": "^11.8.1",
+    "@mui/icons-material": "^5.6.2",
+    "@mui/lab": "^5.0.0-alpha.79",
+    "@mui/material": "^5.6.3",
+    "@mui/x-data-grid": "^5.10.0",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.1.1",
+    "@testing-library/user-event": "^14.1.1",
+    "@types/file-saver": "^2.0.5",
+    "@types/jest": "^27.4.1",
+    "@types/lodash": "^4.14.182",
+    "@types/lodash.isequal": "^4.5.6",
+    "@types/node": "^17.0.30",
+    "@types/pako": "^1.0.3",
+    "@types/react": "^18.0.8",
+    "@types/react-dom": "^18.0.3",
     "copy-to-clipboard": "^3.3.1",
     "file-saver": "^2.0.5",
-    "history": "5",
-    "jszip": "^3.7.1",
+    "history": "^5.3.0",
+    "jszip": "^3.9.1",
     "lodash": "^4.17.21",
     "lodash.isequal": "^4.5.0",
     "notistack": "^2.0.4",
     "pako": "^2.0.4",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "6",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-error-overlay": "6.0.9",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "4.0.3",
-    "typescript": "^4.1.2",
-    "web-vitals": "^1.0.1"
+    "typescript": "^4.6.4",
+    "web-vitals": "^2.1.4"
   },
   "overrides": {
     "postcss": "8.4.5"
@@ -61,11 +66,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "@types/file-saver": "^2.0.4",
-    "@types/lodash": "^4.14.178",
-    "@types/lodash.isequal": "^4.5.5",
-    "@types/pako": "^1.0.3"
   }
 }
