@@ -1,5 +1,5 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
+import {StrictMode} from 'react';
+import {createRoot} from 'react-dom/client';
 import App from './app';
 import {createTheme, CssBaseline, ThemeProvider} from "@mui/material";
 
@@ -44,12 +44,12 @@ const theme = createTheme({
     }
 });
 
-ReactDOM.render(
-    <React.StrictMode>
+const root = document.getElementById('root');
+createRoot(root!).render(
+    <StrictMode>
         <ThemeProvider theme={theme}>
             <CssBaseline/>
             <App/>
         </ThemeProvider>
-    </React.StrictMode>,
-    document.getElementById('root')
+    </StrictMode>
 );
